Add initial structure using poppy flower release.
This is the current flower release based on our current structure.
This effectively provides a structural example.

The files do not have the `.json` extension in an attempt to match the OKAPI request path structure.
This needs to be tested and confirmed.
Ideally, it would be great if the files could have the `.json` extension.