Add message on module descriptor even when not in debug for populate release script.
The Curl command still prints a download progress status to the screen.
This is rather confusing when there is no information describing which Module Descriptor is being fetched.

Add a print message even when not in debug mode to make the download situation more clear.

No message is printed when a file is already downloaded and is skipped while not in debug mode.