Switch to using the snapshot branch for GitHub pages.
The `snapshot` branch will contain all of the module descriptors.
In particular, the `snapshot` releases will be there for the CI/CD integration.

Using a separate branch also provides a separation of concern with regard to the scripts, GitHub Actions, documentation, etc... being on the `main` branch.
This should make the history more navigable.
The `snapshot` branch could also be purged at any point in time should it become too large or problematic without affecting the scripts and GitHub Actions.

This also means that change to the scripts, documentation, or actions will not trigger a rebuild of the GitHub pages.
Only `snapshot` branch changes should trigger this and only Module Descriptors should be stored on the `snapshot` branch (aside from the minimal `README.md` file).