Add Synchronize workflow and script, updating the Gh pages and Populate Release script.
Add `workflow_dispatch:` to enable manual triggering of the GH Pages workflow.

Add Synchronize Snapshot workflow to execute once a day.

Allow for custom inputs, particularly for when manually triggering.

The Synchronize Snapshot workflow is expected to:
  - Checkout the scripts from the main branch.
  - Checkout the `snapshot` branch.
  - Populate the snapshot files using the `populate_release.sh` script.
  - Commit any new files using the `sync_snapshot.sh` script.

This uses the GitHub bot to perform the commits.

This is not yet tested and may require additional changes.

Fix some problems with the variable names in the `populate_release.sh` script where `_RELEASE` is added multiple times.
Refactor the `POPULATE_RELEASE_RELEASE` to `POPULATE_RELEASE_TARGET` to hopefully make it less confusing.
Change the defaults to operate for a `snapshot` release.
Change the environment variable processing to better allow for GitHub Actions where the variables might be defined as an empty string but should be treated as not defined at all.
Remove hard-coded `install.json` file usage, replacing it with the `${file}` variable.
Remove unneeded `-n` parameter from the `grep` command.