Provide Build Latest version links script and workflow.
This script is designed to supplement both the `populate_release.sh` script and the `gh_pages.yml` workflow.

This provides a way to automatically represent the latest version using the `-latest` version tag rather than the explicit version.
This should help improve the CI/CD integration and related tasks.

This uses an incredibly simple version number stripping based on existing data patterns.
There is a good chance, due to its simplicity, than any unusual or different versioning will result in some problems in the generation of the `-latest`.

This provides a ignore setting to allow for not building the latest version link for any particular module.

This fixes some observed mistakes in the `populate_release.sh` script.
Perform some structural re-organization for both the `populate_release.sh` and the `sync_snapshot.sh` scripts.

A stubbed out (empty) `setting/ignore.txt` is provided.