Update the workflows to use inputs, fix typos, and address runner problems.
Fix typo that resulted in the environment variables being pasted in the middle of `..` leading to a failed workflow build.

The self-hosted runner does not clear state between runs.
Manually delete and recreate the `populate` and `scripts` directories at the start.

Add inputs for the workflow trigger events.
Make inputs available to workflow call as well
The inputs default does not appear to be working in a workflow_call.
So, instead, explicitly pass along the `registry_branch` to the `gh_pages.yml` workflow.