Add GitHub Pages generator, improve parallel safety, and miscellaneous fixes.
The primary purpose of this change is to add a GitHub Pages generator.
The previous one being used required Docker and also did not work correctly within our `self-hosted` runner for unknown reasons.
This new script provides more fine-tuned control over how and what is generated.

The GitHub Pages generator is a collection of two parts:
  1. A GitHub Workflow called `gh_pages.yml`.
  2. A Bash script called `build_pages.sh`.

The `gh_pages.yml` Workflow already exists and is modified to utilize the new script, `build_pages.sh`.

All GitHub Workflows have been updated as follows:
  - Improve readability by adding a new line between each run step.
  - Add an Initialize Instance step to improve parallel execution safety in a `self-hosted` runner.
  - Create the working directories at the start and then delete the working directories on success at the end.
  - Have an `id` for each step.
  - Operate each step within the more parallel safe sub-directory.
  - Support the `debug_mode` input variable to allow toggling debug modes for each individual run.
  - Put the inputs in alphabetic mode, except `debug_mode` in the specific case of displaying it as a form input in the GitHub Actions interface.

The `sync_snapshot.yml` Workflow now properly includes the `setting` directory in the sparse checkout.
This was missing before and is essentially a bug.
This was not noticed before because the `settings/ignore.txt` file is currently empty.

Make sure to have git ignore the `work` directory that is now used by the `build_pages.sh` script.

Update the `README.md` as follows:
  - Make GitHub Pages link bold.
  - Add details about the FOLIO Application Generator.
  - Fix heading ordering problem.
  - Add missing example usage for the Build Latest Readme documentation.
  - Add GitHub Workflows Readme documentation, briefly describing each GitHub Workflow.
  - Add Build Pages script documentation to the Readme.

Make scripts more consistent:
  - Don't assign variable to empty using quotes like `debug=""` because assigning empty works like `debug=`.
  - Try to be consistent by wrapping variables in braces `{`, `}` (this is not strictly required in some cases but it helps make the script more compatible with different shells beyond Bash).
  - Print extra new line before the "Done" messages to help make it easier to find in the logs.

Add the `build_pages.sh` script.
See the readme and script documentation for details.
This requires several programs to exist on the runner:
  - bash
  - basename
  - cat
  - cp
  - date
  - grep
  - jq
  - ls
  - mkdir
  - sed

The script is designed with the intention that there might be multiple directories to use.
However, it is not currently designed to handle name conflicts between multiple directories and their respective content.

This script is (and actually all scripts in this repository) are designed to be also run manually on someones local box if they are able.
The following is an example that shows how to build locally and use a local browser to (note some systems need `file://///` instead of `file:///`) view:
```shell
BUILD_PAGES_DEBUG="" BUILD_PAGES_BASE="file:///somewhere" bash script/build_pages.sh release/
```

This script is designed to generate the most simplistic of pages without JavaScript, CSS, or images.
An ordered list is used to help users find or communicate any given line.

A very basic templating system is used to allow for minor customizations.
This templating system also helps keep the HTML out of the script.
The template using very basic regex pattern matching on word boundaries.