Add support for POPULATE_RELEASE_CURL_FAIL and update handle result printing.
The `curl` command does not return a failing state by default when a `404` is received.
Change the setup to pass `--fail` to `curl`, provide custom handling, and make this configurable via `POPULATE_RELEASE_CURL_FAIL`.

The `--fail` mode is called "fail fast" and a `22` is returned as the status code.
This does not catch all cases, so the script is designed to make room for multiple codes to be added in the future if need be.

There are three modes for `POPULATE_RELEASE_CURL_FAIL`.
  1. `fail`: When curl fails with `--fail` passed, then exit the script on error.
  2. `none`: The default curl behavior is performed. A `404` does not trigger a failure and data is downloaded, even if that data is invalid.
  3. `report`: When curl fails with `--fail` and a `22` status is returned, then report the failure but continue running the script (no files are downloaded).

The default is set to `report`.
This is done because there are several FOLIO modules defined that currently return `404`.
These modules may not be needed at this time.
Known problems at this time are:
  - `folio-module-sidecar-2.1.0-SNAPSHOT.172`
  - `mgr-applications-3.1.0-SNAPSHOT.149`
  - `mgr-tenant-entitlements-3.1.0-SNAPSHOT.177`
  - `mgr-tenants-3.1.0-SNAPSHOT.120`

The output gets a little messy when `report` mode is used.
Print a new line after each handle result print call.
Make this change across all scripts for consistency reasons.

The "with system code" error messages are now containing the actual error code.
This change is applied across all scripts for consistency reasons.