Prevent missing files from blocking remaining processes for build_latest.sh.
The `build_latest.sh` script is failing because of:
```
ERROR: Cannot find the version file in eureka-platform.json for folio-module-sidecar to link to: release/snapshot/folio-module-sidecar-3.1.0-SNAPSHOT.176 .
```

This module is not fetchable at this time using the upstream registry.

Add an option, `BUILD_LATEST_SKIP_NOT_FOUND`, to just skip when the file is missing rather than completely fail.
This allows for the everything else that has been fetched to still be published to the registry.
The skipped modules can be later investigated.

The `populate_release.sh` script is printing the curl progress logs by default but the individual file being processed is not printed.
Print a message about the file.
This should help when debugging cases for why `build_latest.sh` failed.