Add Populate Node script and minor changes in scripts and workflows.
Place the environment variables on their own lines when running scripts to make them easier to read.

Add the `populate_node.sh` script and the `workspace` directory.
- This now requires `stripes-cli` to be installed in the runner.
- This generates an `npm.json` file that is intended to be structured the same as the `install.json` file.
- This avoids appending to the `install.json` file to maintain separation of concerns and to avoid `jq` bugs (The `jq` does not handle common command line behaviors and designs very well and requires work-arounds).
- This `npm.json` file gets re-created on every script run (which also helps avoid `jq` problems).
- Try to suppress `yarn` messages when not in debug mode (the warnings aren't easily suppressed unfortunately).
- The module descriptor is generated using the `stripes-cli` which is then copied into the release directory.
- The generated module descriptor is then used to populate the `npm.json` file using the specific version from the `package.json` file.

The `workspace` directory is used to help facilitate downloading the needed packages.
I have been unable to find a way to use the registry to download the package from the NPM registry as the package itself (similar to a git clone).
Instead, the script is forced to add the packages as a dependency to the workspace.
The script then must change into the appropriate `node_modules` sub-directory to process and build the module descriptor.
Using GitHub is out of the question because it fails to store the very specific NPM version number that is needed (such as `folio_authorization-roles-2.0.109900000000137`).

Rename `populate_snapshot` to `populate_release` to be more accurate and consistent with the name.

Make the `DEBUG` parameters be more tolerant of leading and trailing spaces in the debug values for all scripts.
Fix alphabetical ordering issue of some of the `_handle_result()` Bash functions.

Remove the `# Otherwise, any non-empty...` comment to simplify the script documentation.

Add single quotes around inputs in `run-name` to help make debugging easier (the cron runs appear to have a default parameter problem).

Assure that `stripes-cli` is installed when running the workflow.