Work-around GitHub insufficiency regarding defaults during scheduler run.
GitHub fails to honor the input defaults (or fails to provide input defaults support) for scheduler runs.
This insufficiency can either be worked around by mapping the inputs to the outputs and applying a default or by just appending ` || 'some_string'` to every variable usage.

Applying the ` || 'some_string'` logic currently makes more sense because there are a small number of use cases of the input variables.

I would delete the `default` entirely, but this `default` is used in the form input on the GUI side.

This can only be truly tested on a cron run and therefore must be merged before behavior works as intended.

According to the GitHub documentation equality comparisons on an empty string returns `0`.
The use of the `||` (or) operator should work here.

see: https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/evaluate-expressions-in-workflows-and-actions#operators
see: https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/evaluate-expressions-in-workflows-and-actions#functions