Documentation improvements and fix some inconsistencies in the scripts.
Documentation Improvements:
- Add a navigation section due to how large the `README.md` file is.
- Add a note about the `npm-folioci` registry.
- Fix the location of the `Populate via Branches and Commit Hashes` section.
- Add a `Design and Methodology` section along with several sub-sections.
- Add missing `Synchronize Snapshot` section.
- Add `Workflow` to the name of the `GitHub Workflows` sub-sections to help avoid naming conflicts and potential confusion.
- Add the environment variables for the scripts using a table.
- Fix mistake where the `Example usage` shell script is incorrectly swapped with another section.
- Incidental spelling, styling, and grammar mistakes that were noticed are fixed.

Changes to Scripts:
- Relocate the declaration of the `local` variables to be after the `result` variable check (helps avoid adding variables to the stack when not needed).
- Add surrounding `${}` where they have been found to be missing.
- Remove the `ignore.txt` logic and `setting` directory.
- Remove the command line argument parameters support because this is no longer needed.
- Remove the environment variable documentation at the top of the scripts.

The `ignore.txt` logic ended up not being used and is likely no longer now that the "fail forward" logic is in place.
The final design ended up being entirely focused on environment variables.
This makes the command line arguments more costly to have and maintain so remove support for them entirely.

I don't want to document the environment variables in two different places.
It makes perfect sense to have the documentation of the environment variables with the scripts as that keeps them together.
However, the `README.md` is where most people will look and the advanced display features like a table help make reading this list easier.
Relocate the environment variables out of the scripts and into the `README.md` file.
Now the only place to manage them, for the most part, is in the `README.md` file.
More specific details are still provided in the individual scripts.