Fix sed error about "Argument list too long".
The HTML markup that is generated is being passed for every file in a single `sed` statement.
This ended up being way too long.

Shorten this process by processing the template one item at a time, continuously updating the index file in the loop.
This results in a much shorter command line that is less likely to reach the max parameter length limit.

Restructure the code accordingly to accommodate this structural design change.
The `sed` statement is broken up into multiple functions to avoid having to repeat the same sed code in multiple places.
A finalize statement is provided to remove the final `_REPLACE_SECTION_SNIPPET_` and any remaining `_REPLACE_EOL_` values.

The main `index.html` does not have this per-item `sed` structure as the directories there are expected to be a small number.
There may be a case for making the changes for that as well and breaking it up to be processed in a per item basis.