Utilize find and sort piped to while read loop instead of for loop for large file list.
There is a very large limit on how many parameters the `for i in ...` type of loops can handle.
This concern may be alleviated by utilizing a `find` command to programmatically operate on each file in a directory.

The `find` command is able to operate on files one by one.
The `find` command is set to minimum and maximum depths of 1 to match the behavior of the `for i in ...` behavior.
The `find` command does not operate in alphabetic order and needs to have its results piped to the `sort` command.
The `sort` command can then be piped do a `while` `read` loop.
The `while` `read` loop does not preserve the variable changes because it operates in a sub-shell.
Set `shopt -s lastpipe` in the `main()` to ensure that the `while` `read` loop does not operate in a sub-shell.

Relocate the `build_page_operate_sources_process_index_template_item()` to after the empty check so that empty directories do not result in a link to an empty directory on generation.