Improve error reporting.
Have the system code always be reported when using the `*_handle_result()` functions.
This avoids the need of having to add the `(with system code ${result})` in many places.

Always have the `*_handle_result()` functions provide utilize the `${p_e}` error.
This avoids the need of having to add the `${p_e}` everywhere.

Add the period at the end of the `*_handle_result()` functions to properly complete the sentence.
This avoids needing to add the `.` in every string passed to the `*_handle_result()` functions.