Remove dependency on cat and suppress jq errors when using redirect.
It turns out that the `cat` program is entirely unnecessary with bash.

The `<` can be used to redirect data into the `jq` program.

Surprisingly, the `<` can also be used in variable assignment, like this:
```shell
  local my_var=$(< data_to_pipe)
```
The `$(<)` notation is awkward looking but it works.

Use the `&>` instead of `>>` when redirecting to `/dev/null`.
When JSON debugging is disabled, then the `jq` specific errors should be suppressed.
If regular debugging is enabled but JSON debugging is disabled, then continue to use `>>` to still suppress normal `jq` output while still printing the error.