Add sunflower release, R1 2025 (no tag yet, via branch only).
This adds the R1 2025 sunflower release based on the current state of the branch at this time.
The release is subject to change due to lacking an official tag.

The following command is used to generate this:
```sh
POPULATE_RELEASE_REPOSITORY_PART="heads" POPULATE_RELEASE_FLOWER="sunflower" POPULATE_RELEASE_TAG="R1-2025" POPULATE_RELEASE_FILES="install.json" bash script/populate_release.sh
```

Note that `POPULATE_RELEASE_FILES="install.json"` is specified here.
This is because the `eureka-platform.json` is not present in this repository at this time.