Ensure keys with null values are deleted.
The script used to do this due to the presence of `del(f)` in the `jq_instruct` instruction.
Some number of iterations on those instructions resulted in the `del(f)` now longer working as expected.

Add an additional explicit `null` deletion at the end of the instruction to ensure that keys with `null` values are deleted.