Redirect to normal output to ${null} in the specific case of fer-generate_yaml() (another).
This handles a second case where JSON verify is printing when printing is not needed.

The `jq` will print to stdout.
In the specific case of `build_depls_verify_json()`, that output is not needed and is not stored in a variable.
This results in the output printing to the screen, cluttering the screen.
This behavior should only be performed when debugging is enabled.