Implement build_launches.sh script.
This `build_launches.sh` script is intended to merge relevant module descriptor data for use by the `build_deployments.sh`.

The `build_deployments.sh` is updated to process this "launch" data.

The "launch" data is intended to provide the information given by the module developers.
These are applied before most of the local variables to better allow for custom local changes to be performed.
Anything that does not have a custom local change will then utilize the information provided by the module developers.

The generated "launch" files are customizable via a configuration file called `instructions.json`.
The default `instructions.json` pulls in the environment variables and the port information.

There are only two types of customizable processing type supported: `field` and `container_port`.
See the updated `README.md` for further details.

The `jq_` prefix is used in some places that is confusing.
The `jq_` prefix is generally intended to be used for `jq` queries but it has been used for some `jq` results.
Rename these particular cases to be `json_`.

Note that the plural word 'launches' is chosen for lack of a better name to directly refer to the Launch Descriptors.