The app.json must contain an array within items to be a valid fleet manifest and simplify templates.
A single item without the `items` array setup is valid for a fleet manifest.
A top level array of items, however, is not.
Instead, the fleet manifest array must be nested within an `items` object containing the array of items..

Change the design to handle this.
This nested structure increases the complexity.
The YAML format is not safe to use in bash variables so create and use a temporary JSON file for the processing.
At the end of the processed perform one single JSON to YAML conversion.

A new `base.json` file is used to contain the initial empty `items` array structure.

Some testing of the data has been performed and I have been able to identify several properties that do not need to be set.
This is great news as it allows for reducing the size of the templates.

The `namespace` has been discovered to be needed and is now added.