Update templates, add {global:namespace} support.
Add new environment variable `BUILD_DEPLOY_NAMESPACE` to allow for dynamically specifying the namespace via the `{global:namespace}` substitution.

Make changes to the input templates by simplifying the files.
Cross reference with the launches files.
There appears to be a lot of missing information in the launches files making it unclear what should really be configured.

Make sure anything needing the namespace is updated, such as the kong URLs.

Make some of the environment settings more consisent for now.
This is likely to change as needed in the future.
The Max ram percentage is being set to `80%` in general because each docker container should only do one task and that RAM should be dedicated to that task.
The `80%` is being chosen as a conservative decision and could easily be increased as needed.

Reduce the default limit to 1 GB.

There are a lot of template values that are still not well figured out.
I expect there to be a lot more iterations of the templates until we figure out good sane defaults.

The `[JAVA_OPTIONS_MAP]` is changed to `[JAVA_OPTIONS]` for consistency reasons.
The pre-existing `[JAVA_OPTIONS]` is changed to `[JAVA_OPTIONS_VALUE]`.