Improve performance by caching in variables during expansion operations.
The expansion operations have been doing the following inefficient processes:
  - Looping over loading each variable for every field being processed for every pass being made.
  - Reading and writing the results to the file for each operation.

This changes the logic to pre-load the variables once at the start of the script using bash associative arrays.
Keep the expand operations in a variable until the end of the expansion operations in which case the JSON is written to a file.