Provide scripts and processes to identify and designate repository to use.
Some versions exist in the `folioci` repository and some do not.
Others exist in the `folioorg` repository and some do not.

This provides scripts and updates to existings scripts to operate based on pre-configured repositories.

The location of packages must be determined and these locations are stored in a configuration file.
The configuration files are of this format:
```json
{
  "id": "edge-caiasoft-2.5.0-SNAPSHOT.54",
  "repository": "folioci"
}
```
The `id` is essentially the module name and version.
The `repository` is the distinct name as specified in `template/location/repositories.json`.

The `template/location/repositories.json` is a JSON array of repository objects.
The following is an example format, showing only a single repository:
```json
[
  {
    "id": "folioci",
    "name": "folioci",
    "auth": {
      "url": "https://auth.docker.io/",
      "registry": "registry.docker.io"
    },
    "request": {
      "url": "https://registry-1.docker.io/"
    }
  }
]
```

The `id` is an identifier for this specific repository.
The `name` is the name used in the location file to represent this repository (such as `folioci` in `https://hub.docker.com/r/folioci/mod-circulation`).
The `auth` is broken up into two parts, a `url` and a `registry`.
The `auth.url` is optional; This is used to perform GET requests to obtain an authentication token as an anonymous user.
The `auth.registry` is optional and is only used if `auth.url` is used; provides information needed for requesting the authentication token.
The `request` is broken up into one part, a `url`.
The `request.url` is the URL used to perform the HEAD request used to identify the presence of a package and version at the named repository.

There are notable concerns with making requests to the remote servers regarding rate limiting and request limiting.
See the comments in the `build_location.sh` script for details.