Add support for services and update template files.
The `king: service` type needs to also be built.
Extend the design to enable support a `service`.
The mapping file now has an object with the `deploy` and `service` keys.
The `deploy` key represents the original behavior with the original mapping file.
The `service` is the new key.

Centralize more of the side car settings.
Many of the side car settings can be removed out of the `JAVA_OPTIONS` and be specified via the environment variables.

Disable used of the probes for the side car for now (should change in the future).

Make sure ports are accurate.
Set the default ports to use `8081` and any exception cases must have manual port assignments.

Increase the values for the liveness and readiness probes.
The previous values are too low and many slow starting pods are being forced to restart before they are ready.

One notable proble is that liquibase is locking the tables which then causes failures.
Disable liquibase (and possibly flywy) on these images on start using environment variables.
- Set `SPRING_LIQUIBASE_ENABLED` to `false`.
- Set `SPRING_FLYWAY_ENABLED` to `false`.

Had to manually delete the lock file in the DB:
```
folio_modules=# select  * from  pubsub_config.databasechangeloglock ;
 id | locked |        lockgranted         |                 lockedby
----+--------+----------------------------+-------------------------------------------
  1 | t      | 2025-06-02 19:53:17.302663 | mod-pubsub-5fd99858d7-5z2g6 (10.42.9.244)
(1 row)

folio_modules=# delete from  pubsub_config.databasechangeloglock ;
```