Support "app", tweak use of labels, and disable liveness probe.
Change `BUILD_DEPLOY_OUTPUT_FILE` to `BUILD_DEPLOY_APP` and expose the app as `{global:app}`.
This allows for adding a label with the app name, such as `app-acquisitions`.
The file name of the generated YAML output file will be the app name.

Utilize the following labels:
- `app`: The application name.
- `app.kubernetes.io/name`: The module name. This is a special kubernetes field.
- `app.kubernetes.io/instance`: The unique ID of the module instance. This is a special kubernetes field.
- `app.kubernetes.io/part-of`: An association, always set to `FOLIO`. This is a special kubernetes field.
- `app.kubernetes.io/version`: The module version. This is a special kubernetes field.
- `environment`: The environment, such as `scratch`.

There are problems with the probes so liveness is temporarily disabled.
Increase the probe failure thresholds to 10.

see: https://kubernetes.io/docs/concepts/overview/working-with-objects/common-labels/
see: https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/