Empty array results in invalid length assignment.
Using the `let=` bash assignment directly on a `jq` response is not good.
The `jq` will return either the string `null` or an empty string.

Using `let=` with a value of `null` produces an error.

Move the `let=` assignment to a condition that checks the response.
If the `result` is either `null` or an empty string then explicitly set the value to `0`.
Otherwise, assign the `result` as is using `let=`.