Address security concern regarding GitHub Action script injections.
Github variables must not directly be used in inline scripts due to how GitHub performs substitution.
Instead, these should be assigned to an environment variable and then call the environment variable (with double quotes around the environment variable use).

see: https://docs.github.com/en/actions/security-for-github-actions/security-guides/security-hardening-for-github-actions#understanding-the-risk-of-script-injections