Improve bash security by locking down IFS changes.
The `IFS` environment variable is a special variable that can be used to change how variable expansion works.
This can be used to break scripts or even perform exploits.

Lock this down by clearing out `IFS` before operating anything in the script.
A local variable scope is defined to prevent this change from bubbling up and affecting the callers environment.

Example script showing the behavior of `IFS`:

```shell
IFS=":"
echo ${PATH}
```

Paths are normally separated by a `:` but are not separated by a space ` `.