Add Application Descriptor Generator.
Many of the scripts can and need to be run in a non-Eureka/OKAPI-only environment.
At least, until one fully migrates to Eureka.
This generator is useful as a transitory script for using the newer processes in an older OKAPI environment.

The Application Descriptor is generated for Eureka using projects like the `app-platform-complete` and `folio-application-generator`.
Many of these, however, are Eureka focused.

The Application Descriptor Generator provides a way to create a basic Application Descriptor using an `install.json` without needing a full project, such as `app-platform-complete`.

This script is not intended to be used to generate a complete Application Descriptor.
Instead, this script constructs the bare minimum parts of an Application Descriptor needed to execute the Build Module Discovery scripts in this repository.