Do not check return status of grep.
The `grep` documentation states the following:
```
Exit status is 0 if any line is selected, 1 otherwise;
if any error occurs and -q is not given, the exit status is 2.
```

Therefore, an exist status of `1` is not an error.

Fix a false positive in `build_depls_expand_file_load_template_maps_pcre_match_query()`.