Implement Module and deployment descriptor generating script.
See `README.md` for some details on how the script works.

Notes about some of the default settings:

The `edge-connexion` lacks a discovery descriptor and must be skipped.
The `mod-graphql` uses custom names for the descriptors and those descriptors are pre-created.

The `mod-camunda` and `mod-workflow` are used as a functioning example in the default `maps.json`.

The UI modules don't use deployment descriptors and must be skipped.

The `mod-z3950` is missing or requires special handling and is therefore omitted.
The `okapi` module is omitted.

Some of the stripes modules, such as `folio_stripes-authority-components`, `folio_stripes-core`, and `folio_stripes-smart-components` must be renamed.
Their projects lack the traditional `ui-` prefix.
It might be better to have these be skipped as they are components or core dependencies.