Update code, fixing issues and improving consistency.
The `BUILD_APP_DESCRIPTOR_RESTRICT_TO` should be allowed to be allowed to be empty if restriction is not needed.

The variable parameters passed to functions and programs must all be quoted to prevent problems with passing arguments.
The variables can expand into multiple arguments due to possible spaces in the variable and can have undesired effects.
Only exceptional situations where this behavior is expected and intended are not quoted, such as with `${debug}`.
Non-variables are not required to be quoted (should they be for consistency?).

Make sure local variable assignments all have the `=` sign.

Make sure all `*_load_environment` functions pass the arguments (without quotes to expand).

The `pcre_total` should be a number, so use `local -i` instead of `local`.

Check the `flower` variable for invalid characters in all scripts that utilize this variable.

Remove the double quote usage for associative arrays to make the code more consistent.
The double quotes can cause problems when nested inside double quotes.
Rather than escaping, just remove the quotes entirely given that bash appears to allow this.
This helps reduce the chance of mistakes.

Replace all echo pipes, such as `echo "a" | grep -sho "a"`, with the three less than sign pipe, such as `grep -sho "a" <<< "a"`.
The `<<<` pipe does not do additional processing and is a little simpler than `echo |`.
This also results in the import command being executed to be on the left, thereby making the command being run more obvious.

Wrap file paths in quotes to be on the safe side.
The file paths themselves can have spaces.

Try to keep the `${debug}` and similar parameters as far left in the parameters where possible.